Add in support to unlike pages that were liked within the max/min date range. Also add in support to untag photos - at least those that show up on the photos feed.  Chances are there's lots missing there, so this needs a bit more work to be complete.
Change the behaviour so that rather than passing in a single function (action_f)
to perform_xpath the caller passes in a dictionary of functions.  Then change
things so that the list of tuples can contain a function name.  This makes
it so the caller can override the default stuff, and also add their own, while
allowing each xpath match to perform a different thing.  Add a new hover
operation to hover rather than click (needed for untagging.)